requirements.txt
dj-database-url==0.5.0
Django==3.1.6
django-heroku==0.3.1

decorator==4.4.0

django-allauth==0.42.0
django-countries==7.0
django-crispy-forms==1.10.0
django-filebrowser-no-grappelli==3.8.0
django-filter==2.3.0
django-tinymce==3.2.0
djangorestframework==3.11.1
docutils==0.15.2
django-heroku==0.3.1
django-rq-wrapper==3.1
gunicorn==20.0.4
dj3-cloudinary-storage==0.0.5



oauthlib==2.1.0

Pillow==5.4.1

psycopg2-binary==2.9.1

cloudinary==1.25.0


SecretStorage==2.3.1

whitenoise==5.2.